+++ Logic/VacationLogic.ts
@@ -1,7 +1,9 @@
 import dal_mysql from "../Utils/dal_mysql";
-import { OkPacket } from "mysql";
+import {OkPacket} from "mysql";
 import Vacation from './../Models/Vacation';
 import randomUtils from '../Utils/randonUtils';
+import {MOCK_DATA, mockVacations} from "../mockData";
+import followerLogic from './FollowerLogic'
 
 const createVacationsTable = async () => {
   const SQLcommand = `
@@ -41,6 +43,10 @@ const dropVacationsTable = async () => {
 };
 
 const addVacation = async (newVacation: Vacation) => {
+  if (MOCK_DATA) {
+    mockVacations.push(newVacation);
+    return newVacation.vacationId
+  } else {
     const SQLcommand = `
   INSERT INTO project3.vacations 
   (destination,description,startDate,endDate,price,image) 
@@ -55,44 +61,49 @@ const addVacation = async (newVacation: Vacation) => {
 
     const result: OkPacket = await dal_mysql.execute(SQLcommand);
     return result.insertId;
+  }
 };
 
 const deleteVacation = async (vacationId: number) => {
-  const SQLcommand = `DELETE FROM project3.vacations WHERE vacationId=${vacationId}`;
-  return dal_mysql.execute(SQLcommand).then(() => true).catch(() => false);
+  if (MOCK_DATA) {
+    const index: number = mockVacations.findIndex(vacation => vacation.vacationId === vacationId);
+    mockVacations.splice(index, 1);
+    return true;
+  } else {
+    const SQLcommand = `DELETE FROM project3.vacations WHERE vacationId=${vacationId}`;
+    return dal_mysql.execute(SQLcommand).then(() => true).catch(() => false);
+  }
+
 };
 
-const getAllVacations = async (userId: number) => {
-  /*const SQLcommand = `SELECT
-  v.vacationId,v.destination,v.description,v.startDate,v.endDate,v.price,v.image,
-  COUNT(f.userId) AS followersCount,
-    CASE f.userId
-      WHEN ${userId} THEN true
-      ELSE false
-    END AS isFollow
-FROM
-  vacations v
-  LEFT JOIN followers f 
-  ON v.vacationId = f.vacationId
-GROUP BY
-  v.vacationId, isFollow`;*/
-  const SQLcommand = "SELECT * FROM project3.vacations";
-  /*const SQLcommandCombined = `SELECT
-  innerTable.vacationId,innerTable.destination,innerTable.description,innerTable.startDate,innerTable.endDate,innerTable.price,innerTable.image,
-  SUM(innerTable.followersCount) as followersCount,
-    SUM(innerTable.isFollow) as isFollow
-    FROM (${SQLcommand}) innerTable group by innerTable.vacationId;`;*/
-  return await dal_mysql.execute(SQLcommand);
+const getAllVacations = async () => {
+  if (MOCK_DATA) {
+    return mockVacations;
+  } else {
+    const SQLcommand = "SELECT * FROM project3.vacations";
+    return await dal_mysql.execute(SQLcommand);
+  }
 };
 
 const getSingleVacation = async (vacationId: number): Promise<Vacation> => {
-  const SQLcommand = `SELECT * FROM project3.vacations WHERE vacationId=${vacationId}`;
-  const singleVacation = await dal_mysql.execute(SQLcommand);
-  return singleVacation[0];
+  if (MOCK_DATA) {
+    return mockVacations.find(vacation => vacation.vacationId === vacationId);
+  } else {
+    const SQLcommand = `SELECT * FROM project3.vacations WHERE vacationId=${vacationId}`;
+    const singleVacation = await dal_mysql.execute(SQLcommand);
+    return singleVacation[0];
+  }
+
 };
 
 const updateVacation = async (vacation: Vacation) => {
-  const SQLcommand = `UPDATE project3.vacations
+  if (MOCK_DATA) {
+    const index: number = mockVacations.findIndex(vacation => vacation.vacationId === vacation.vacationId);
+    mockVacations[index] = vacation;
+    return true;
+
+  } else {
+    const SQLcommand = `UPDATE project3.vacations
   SET destination = ${vacation.destination}, 
   description = ${vacation.description}, 
   startDate = ${vacation.startDate}, 
@@ -100,8 +111,21 @@ const updateVacation = async (vacation: Vacation) => {
   price = ${vacation.price}, 
   image = ${vacation.image} 
   WHERE vacationId=${vacation.vacationId};`;
-  return dal_mysql.execute(SQLcommand).then(() => true).catch(() => false);
+    return dal_mysql.execute(SQLcommand).then(() => true).catch(() => false);
+  }
+
 };
+
+
+const getFollowedVacationsOfUser = async (userId: number) => {
+  const vacationIds = await followerLogic.getVacationFollowsOfUser(userId)
+    if (MOCK_DATA) {
+        return mockVacations.filter(vacation => vacationIds.includes(vacation.vacationId));
+    } else {
+        const SQLcommand = `SELECT * FROM project3.vacations WHERE vacationId IN (${vacationIds.join(',')})`;
+        return await dal_mysql.execute(SQLcommand);
+    }
+}
 
   export default {
     dropVacationsTable,
@@ -111,4 +135,5 @@ const updateVacation = async (vacation: Vacation) => {
     getAllVacations,
     updateVacation,
     getSingleVacation,
+    getFollowedVacationsOfUser,
   };
